[FEAT] 1차 병합, 인덱스 타입 Long으로 변경

+++ src/main/java/com/spoony/spoony_server/domain/location/entity/LocationDongEntity.java
@@ -11,14 +11,14 @@ import lombok.NoArgsConstructor;
 public class LocationDongEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer guId;
+    private Long guId;
     private String guName;
     private String guAddress;
     private Double latitude;
     private Double longitude;
 
     @Builder
-    public LocationDongEntity(String guName, Integer guId, String guAddress, Double latitude, Double longitude) {
+    public LocationDongEntity(String guName, Long guId, String guAddress, Double latitude, Double longitude) {
         this.guName = guName;
         this.guId = guId;
         this.guAddress = guAddress;
