+++ src/main/java/com/spoony/spoony_server/domain/place/repository/PlaceRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface PlaceRepository extends JpaRepository<PlaceEntity, Long> {
-    boolean existsByPlaceIdInAndLatitudeAndLongitude(List<Integer> placeIds, Double latitude, Double longitude);
+    boolean existsByPlaceIdInAndLatitudeAndLongitude(List<Long> placeIds, Double latitude, Double longitude);
 }
