+++ src/main/java/com/spoony/spoony_server/domain/post/repository/PostRepository.java
@@ -9,11 +9,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PostRepository extends JpaRepository<PostEntity, Integer> {
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
 
-    Optional<MenuEntity> findMenuByPostId(Integer postId);
+    Optional<MenuEntity> findMenuByPostId(Long postId);
 
-    Integer countByPostId(Integer postId);
+    Long countByPostId(Long postId);
 
     List<PostEntity> findByUser_UserId(Long userId);
 }
